23832 - fix: update height of the content. refactor to make it more easily composable.
