+++ btr-web/btr-layouts/layouts/business.vue
@@ -1,6 +1,6 @@
 <template>
   <div data-cy="business-layout">
-    <BcrosHeader />
+    <BcrosCustomHeader />
     <BcrosBreadcrumb v-if="crumbConstructors.length > 0" :crumb-constructors="crumbConstructors" />
     <BcrosBusinessDetails />
     <div class="mx-auto px-4 w-full max-w-bcroslg flex">
@@ -29,7 +29,7 @@
       :left-button-constructors="leftButtonConstructors"
       :right-button-constructors="rightButtonConstructors"
     />
-    <BcrosFooter :app-version="version" />
+    <BcrosCustomFooter :app-version="version" />
   </div>
 </template>
