+++ btr-web/btr-layouts/tests/unit/components/bcros/Footer.spec.ts
@@ -1,17 +1,17 @@
 import { expect, describe, test } from 'vitest'
 import { mountSuspended } from '@nuxt/test-utils/runtime'
 
-import { BcrosFooter } from '#components'
+import { BcrosCustomFooter } from '#components'
 
 describe('Tests for Footer.vue', () => {
   test('UTooltip component rendered', async () => {
-    const wrapper = await mountSuspended(BcrosFooter, { props: { appVersion: '1' } })
+    const wrapper = await mountSuspended(BcrosCustomFooter, { props: { appVersion: '1' } })
     const tooltip = wrapper.find('#footer-tooltip')
     expect(tooltip.exists()).toBeTruthy()
   })
 
   test('footer nav links rendered', async () => {
-    const wrapper = await mountSuspended(BcrosFooter, { props: { appVersion: '1' } })
+    const wrapper = await mountSuspended(BcrosCustomFooter, { props: { appVersion: '1' } })
 
     const expectedTexts = ['Home', 'Release Notes', 'Disclaimer', 'Privacy', 'Accessibility', 'Copyright']
     const expectedHrefs = [
