+++ btr-web/btr-main-app/enums/route-name-e.ts
@@ -3,6 +3,7 @@ export enum RouteNameE {
   MY_REG_DETAILS = 'my-registries-details',
   REVIEW_CONFIRM = 'review-confirm',
   REQUEST_OMIT = 'request-to-omit',
+  REQUEST_OMIT_SUBMITTED = 'request-to-omit-submitted',
   STAFF_SI_DASH = 'staff-si-dashboard',
   STAFF_REQ_VIEW = 'staff-request-view'
 }
